Port man page to mkdocs
This patch moves the documentation from man pages to
[MkDocs](http://mkdocs.org).

Install MkDocs with `sudo pip mkdocs` and run `mkdocs server` in the `man`
directory, then you can browse the documentation at http://localhost:8000.

This is a proof of concept. It keeps the content of the man page, reuses the
existing style sheet, and adds a navigation menu.

To make it ready for production use we would at least need to:

* Find a solution for generating the scopes documentation
* Put documentation in a separate directory (`manual` maybe?)
* Add a new man page which contains some general info and points to the mkdocs
  online version
* Fix the rendering errors in the titles of the commands
* Adapt the Rake tasks to generate the man page
* Adapt the release script to push the generated documentation
* Review the formatting and structure of the documentation