Update tests for new os version and name format
At the moment our display_name for Unknown oses which is used in
several places is put together like this:
"#{name} #{version} (#{architecture})"

The result for Ubuntu for example is:
Ubuntu 14.04.3 LTS 14.04.3 LTS, Trusty Tahr (x86_64)

Or Fedora 23 Server Edition:
Fedora 23 (Server Edition) 23 (Server Edition) (x86_64)

For all newer oses we parse the os-release.
Instead of the PRETTY_NAME and VERSION we should use NAME and
VERSION_ID.

This will prevent redundancy and give a proper output unless
we think that the variant_name is important.